Data-Sender/requirements.txt
requests
google-cloud-pubsub==2.23.0